mnCharSel_CursorThink: index the door with the plain loop counter
Retail multiplies the counter directly; the narrowing cast emitted an
extsb it does not have. It scored well before the loop's break was
fixed, which was masking the real shape.

+++ src/melee/mn/mncharsel.c
@@ -2768,7 +2768,7 @@ void mnCharSel_CursorThink(HSD_GObj* gobj)
                                             {
                                                 u8 new_kind;
                                                 pk = &mnCharSel_803F0DFC
-                                                          .doors[(s8) di]
+                                                          .doors[di]
                                                           .p_kind;
                                                 new_kind = *pk + 1;
                                                 *pk = new_kind;
